rename salt.utils.socket_util to salt.utils.network
This change reflects the changed role of this module, as a more general
collection of networking functions.
