+++ salt/modules/network.py
@@ -32,7 +32,7 @@ def ping(host):
 
         salt '*' network.ping archlinux.org
     '''
-    cmd = 'ping -c 4 {0}'.format(salt.utils.socket_util.sanitize_host(host))
+    cmd = 'ping -c 4 {0}'.format(salt.utils.network.sanitize_host(host))
     return __salt__['cmd.run'](cmd)
 
 
@@ -83,7 +83,7 @@ def traceroute(host):
         salt '*' network.traceroute archlinux.org
     '''
     ret = []
-    cmd = 'traceroute {0}'.format(salt.utils.socket_util.sanitize_host(host))
+    cmd = 'traceroute {0}'.format(salt.utils.network.sanitize_host(host))
     out = __salt__['cmd.run'](cmd)
 
     # Parse version of traceroute
@@ -138,7 +138,7 @@ def dig(host):
 
         salt '*' network.dig archlinux.org
     '''
-    cmd = 'dig {0}'.format(salt.utils.socket_util.sanitize_host(host))
+    cmd = 'dig {0}'.format(salt.utils.network.sanitize_host(host))
     return __salt__['cmd.run'](cmd)
 
 
@@ -168,7 +168,7 @@ def interfaces():
 
         salt '*' network.interfaces
     '''
-    return salt.utils.socket_util.interfaces()
+    return salt.utils.network.interfaces()
 
 
 def hwaddr(iface):
@@ -179,7 +179,7 @@ def hwaddr(iface):
 
         salt '*' network.hwaddr eth0
     '''
-    return salt.utils.socket_util.hwaddr(iface)
+    return salt.utils.network.hwaddr(iface)
 
 
 def subnets():
@@ -190,7 +190,7 @@ def subnets():
 
         salt '*' network.subnets
     '''
-    return salt.utils.socket_util.subnets()
+    return salt.utils.network.subnets()
 
 
 def in_subnet(cidr):
@@ -201,7 +201,7 @@ def in_subnet(cidr):
 
         salt '*' network.in_subnet 10.0.0.0/16
     '''
-    return salt.utils.socket_util.in_subnet(cidr)
+    return salt.utils.network.in_subnet(cidr)
 
 
 def ip_addrs(interface=None, include_loopback=False):
@@ -214,8 +214,8 @@ def ip_addrs(interface=None, include_loopback=False):
 
         salt '*' network.ip_addrs
     '''
-    return salt.utils.socket_util.ip_addrs(interface=interface,
-                                           include_loopback=include_loopback)
+    return salt.utils.network.ip_addrs(interface=interface,
+                                       include_loopback=include_loopback)
 
 
 def ip_addrs6(interface=None, include_loopback=False):
@@ -228,5 +228,5 @@ def ip_addrs6(interface=None, include_loopback=False):
 
         salt '*' network.ip_addrs6
     '''
-    return salt.utils.socket_util.ip_addrs6(interface=interface,
-                                            include_loopback=include_loopback)
+    return salt.utils.network.ip_addrs6(interface=interface,
+                                        include_loopback=include_loopback)
