+++ salt/utils/network.py
